requirements.txt
streamlit>=1.24
langchain-core>=1.2.5
langchain-community>=0.4.1
langchain-text-splitters>=1.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-docx>=0.8.11
docx2txt>=0.8.0
pypdf>=3.0.0
unstructured>=0.6.9
beautifulsoup4>=4.12.2
lxml>=4.9.2
pandas>=2.0.0
python-pptx>=0.6.21
requests>=2.28.0
tf-keras>=2.12.0
openai>=0.27.0
openpyxl>=3.0.10

